+++ include/awkward/typedbuilder/ListArrayBuilder.h
@@ -9,6 +9,7 @@ namespace awkward {
 
   class ListForm;
   using ListFormPtr = std::shared_ptr<ListForm>;
+  using FormBuilderPtr = std::shared_ptr<FormBuilder>;
 
   /// @class ListArrayBuilder
   ///
@@ -24,7 +25,7 @@ namespace awkward {
 
     /// @brief Turns the accumulated data into a Content array.
     const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
+      snapshot(const ForthOutputBufferMap& outputs) const override;
 
     /// @brief
     const FormPtr
