+++ src/libawkward/typedbuilder/FormBuilder.cpp
@@ -1,0 +1,11 @@
+// BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
+
+#define FILENAME(line) FILENAME_FOR_EXCEPTIONS("src/libawkward/builder/TypedArrayBuilder.cpp", line)
+
+#include "awkward/typedbuilder/FormBuilder.h"
+
+namespace awkward {
+
+  FormBuilder::~FormBuilder() = default;
+
+}
